# Release v0.4.3.2b 2025-01-05
## New Features
1. make replication shard count agnostic (v0.4.3.0b 2025-12-28)
2. add debug replication mode (v0.4.3.0b 2025-12-28)
3. Improve parser speed on micro-benchmarks (v0.4.3.2b 2025-01-02)
4. calculate default max memory to be 3/4 of physical memory (Release v0.4.3.2b 2025-01-05)

## Fixes
1. fix string push along with `CLIENT INFO` (v0.4.3.0b 2025-12-29)
2. hash test did not test hash (v0.4.3.0b 2025-12-29)
3. remove unused code (v0.4.3.1b 2025-12-30)
4. fix incr/decr for doubles (v0.4.3.1b 2025-12-30)
5. fix memcpy/buffer overflow bug (v0.4.3.2b 2025-12-31)

+++ src/rpc/barch_rpc.h
@@ -27,7 +27,7 @@ enum {
 namespace barch {
 
     template<typename T>
-    bool time_wait(int64_t millis, T&& fwait ) {
+    bool time_wait(uint64_t millis, T&& fwait ) {
         if (fwait()) {
             return true;
         }
@@ -37,7 +37,7 @@ namespace barch {
                 return true;
             }
             auto now = std::chrono::steady_clock::now();
-            if (std::chrono::duration_cast<std::chrono::milliseconds>(now - start).count() > millis) {
+            if (std::chrono::duration_cast<std::chrono::milliseconds>(now - start).count() > (int64_t)millis) {
                 return false;
             }
             std::this_thread::sleep_for(std::chrono::milliseconds(1));
